Move functional utilities out of `H5GroveApi` class
